supplywatch/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
SQLAlchemy==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pydantic-settings==2.9.1
apscheduler==3.10.4
httpx==0.28.1
resend==2.8.0
python-dotenv==1.1.0
pytest==8.3.5
pytest-asyncio==0.26.0
